+++ package.json
@@ -12,6 +12,7 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "@googlemaps/react-wrapper": "^1.1.35",
     "argon2": "^0.29.1",
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
