+++ server/public/styles.css
@@ -1,20 +1,12 @@
 img {
   width: 100%;
   max-height: 100%;
-  object-fit: cover;
+  object-fit: contain;
+  background-color: #e5e5e5;
 }
 
 .image-container {
   height: 300px;
-}
-
-.block-button {
-  width: 100%;
-  background-color: rgb(127, 182, 127);
-  border-radius: 0.3rem;
-  border: none;
-  color: white;
-  height: 2.5rem;
 }
 
 label {
